add checking for existence of order

+++ core/src/main/scala/dao/order/OrderDao.scala
@@ -6,7 +6,7 @@ import zio.{Task, ZLayer}
 
 trait OrderDao {
   def saveOrder(order: Order): Task[Unit]
-  def getOrder(campaignId: Int, orderId: Int): Task[Option[Order]]
+  def getOrder(campaignId: Long, orderId: Long): Task[Option[Order]]
   def getOrdersByStatus(status: Status): Task[List[Order]]
   def changeStatus(orders: List[Order], status: Status): Task[Unit]
   def enrichOrders(enrichedOrders: List[Order]): Task[Unit]
